feat: Refactor auth files
Uses a file structure uses in nest docs and makes more sense as we add further guards.

Note: There's some gross importing of repo modules currently, that'll all be refactored out shortly.
