+++ server/src/@common/decorators/logged-in-user.decorator.ts
@@ -1,7 +1,7 @@
 ﻿import { createParamDecorator, ExecutionContext } from '@nestjs/common';
 
 // This will always be behind auth guard so user should always been in request
-export const LoggedInUser = createParamDecorator((data: string | undefined, ctx: ExecutionContext) => {
-    const request = ctx.switchToHttp().getRequest();
+export const LoggedInUser = createParamDecorator((data: string | undefined, context: ExecutionContext) => {
+    const request = context.switchToHttp().getRequest();
     return data ? request.user[data] : request.user;
 });
