+++ server/src/modules/auth/auth.controller.ts
@@ -2,11 +2,11 @@ import { Req, Res, Controller, Get, Post, HttpException, UseGuards } from '@nest
 import { Request } from 'express';
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import * as passport from 'passport';
-import { Public } from './public.decorator';
+import { Public } from '../../@common/decorators/public.decorator';
 import { AuthService } from './auth.service';
-import { SteamWebAuthGuard } from './steam/steam-web-auth.guard';
+import { SteamWebAuthGuard } from './guard/steam-web-auth.guard';
 import { User } from '@prisma/client';
-import { SteamAuthService } from './steam/steam-auth.service';
+import { SteamAuthService } from './steam-auth.service';
 import { JWTResponseDto } from '../../@common/dto/jwt-response.dto';
 
 @Controller('/auth')
