+++ server/src/modules/auth/strategy/jwt.strategy.ts
@@ -3,11 +3,11 @@ import { PassportStrategy } from '@nestjs/passport';
 import { Injectable } from '@nestjs/common';
 import { JWTPayload } from '../auth.service';
 import { appConfig } from '../../../../config/config';
-import { UsersService } from '../../users/users.service';
+import { UsersRepo } from '../../users/users.repo';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
-    constructor(private readonly usersService: UsersService) {
+    constructor(private readonly userRepo: UsersRepo) {
         super({
             jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
             ignoreExpiration: false,
@@ -19,7 +19,7 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
      * @summary Is this a valid JWT?
      */
     async validate(validationPayload: JWTPayload) {
-        // if its valid then this will hit
-        return this.usersService.Get(validationPayload.id);
+        // TODO: move after repo refactor
+        return this.userRepo.Get(validationPayload.id);
     }
 }
